Refactor Endpoints and Services part 3

+++ back/app/cafe_menu/endpoints.py
@@ -8,6 +8,7 @@ from beanie import PydanticObjectId
 from fastapi import APIRouter, Depends, HTTPException, Path, Query, Request, status
 from fastapi_pagination.ext.beanie import paginate
 from fastapi_pagination.links import Page
+from pymongo.errors import DuplicateKeyError
 
 from app.auth.dependencies import get_current_user
 from app.cafe.models import Role
@@ -21,7 +22,7 @@ from app.cafe_menu.models import (
     MenuItemUpdate,
 )
 from app.cafe_menu.service import MenuService
-from app.models import ErrorResponse
+from app.models import ErrorConflictResponse, ErrorResponse
 from app.service import parse_query_params
 from app.user.models import User
 
@@ -49,8 +50,21 @@ async def create_menu_category(
     current_user: User = Depends(get_current_user),
 ):
     """Create a menu category. (`admin`)"""
-    # TODO: Implement authorization for staff
     cafe = await CafeService.get(slug)
+    if not cafe:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=[{"msg": "A cafe with this slug does not exist."}],
+        )
+
+    category = await MenuService.get_category_by_name(cafe, data.name)
+    if category:
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail=[{"msg": "A category with this name already exists."}],
+        )
+
+    await CafeService.is_authorized_for_cafe_action(cafe, current_user, [Role.ADMIN])
     return await MenuService.create_category(cafe, data)
 
 
@@ -71,8 +85,28 @@ async def update_menu_category(
     current_user: User = Depends(get_current_user),
 ):
     """Update a menu category. (`admin`)"""
-    # TODO: Implement authorization for staff
     cafe = await CafeService.get(slug)
+    if not cafe:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=[{"msg": "A cafe with this slug does not exist."}],
+        )
+
+    category = await MenuService.get_category_by_id(cafe, id)
+    if not category:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=[{"msg": "A category with this ID does not exist."}],
+        )
+
+    category = await MenuService.get_category_by_name(cafe, data.name)
+    if category and category.id != id:
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail=[{"msg": "A category with this name already exists."}],
+        )
+
+    await CafeService.is_authorized_for_cafe_action(cafe, current_user, [Role.ADMIN])
     return await MenuService.update_category(cafe, id, data)
 
 
@@ -90,8 +124,21 @@ async def delete_menu_category(
     current_user: User = Depends(get_current_user),
 ):
     """Delete a menu category. (`admin`)"""
-    # TODO: Implement authorization for staff
     cafe = await CafeService.get(slug)
+    if not cafe:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=[{"msg": "A cafe with this slug does not exist."}],
+        )
+
+    category = await MenuService.get_category_by_id(cafe, id)
+    if not category:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=[{"msg": "A category with this ID does not exist."}],
+        )
+
+    await CafeService.is_authorized_for_cafe_action(cafe, current_user, [Role.ADMIN])
     await MenuService.delete_category(cafe, id)
 
 
@@ -117,7 +164,8 @@ async def get_menu_items(
     cafe = await CafeService.get(slug)
     if not cafe:
         raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Cafe not found"
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=[{"msg": "A cafe with this slug does not exist."}],
         )
 
     filters = parse_query_params(dict(request.query_params))
@@ -134,7 +182,7 @@ async def get_menu_items(
         401: {"model": ErrorResponse},
         403: {"model": ErrorResponse},
         404: {"model": ErrorResponse},
-        409: {"model": ErrorResponse},
+        409: {"model": ErrorConflictResponse},
     },
 )
 async def create_menu_item(
@@ -146,16 +194,32 @@ async def create_menu_item(
     cafe = await CafeService.get(slug)
     if not cafe:
         raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Cafe not found"
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=[{"msg": "A cafe with this slug does not exist."}],
         )
 
+    if data.category_id is not None:
+        category = await MenuService.get_category_by_id(cafe, data.category_id)
+        if not category:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail=[{"msg": "A category with this ID does not exist."}],
+            )
+
+    await CafeService.is_authorized_for_cafe_action(cafe, current_user, [Role.ADMIN])
+
     try:
-        await CafeService.is_authorized_for_cafe_action(
-            cafe, current_user, [Role.ADMIN]
-        )
         return await MenuService.create_item(cafe, data)
-    except ValueError as e:
-        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=str(e))
+    except DuplicateKeyError as e:
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail=[
+                {
+                    "msg": "An item with these fields already exists.",
+                    "fields": list(e.details.get("keyPattern", {}).keys()),
+                }
+            ],
+        )
 
 
 @menu_router.get(
@@ -166,13 +230,22 @@ async def create_menu_item(
     },
 )
 async def get_menu_item(
+    slug: str = Path(..., description="Slug of the cafe"),
     id: PydanticObjectId = Path(..., description="ID of the menu item"),
 ):
     """Get a menu item."""
-    item = await MenuService.get_item(id)
+    cafe = await CafeService.get(slug)
+    if not cafe:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=[{"msg": "A cafe with this slug does not exist."}],
+        )
+
+    item = await MenuService.get_item_by_id_and_cafe_id(id, cafe.id)
     if not item:
         raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Menu item not found."
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=[{"msg": "An item with this ID does not exist."}],
         )
     return item
 
@@ -184,39 +257,54 @@ async def get_menu_item(
         401: {"model": ErrorResponse},
         403: {"model": ErrorResponse},
         404: {"model": ErrorResponse},
-        409: {"model": ErrorResponse},
+        409: {"model": ErrorConflictResponse},
     },
 )
 async def update_menu_item(
     data: MenuItemUpdate,
+    slug: str = Path(..., description="Slug of the cafe"),
     id: PydanticObjectId = Path(..., description="ID of the menu item"),
     current_user: User = Depends(get_current_user),
 ):
     """Update a menu item. (`volunteer`)"""
-    item = await MenuService.get_item(id)
-    if not item:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Menu item not found."
-        )
-
-    cafe = await CafeService.get(item.cafe_id)
+    cafe = await CafeService.get(slug)
     if not cafe:
         raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Cafe not found"
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=[{"msg": "A cafe with this slug does not exist."}],
         )
 
-    try:
-        await CafeService.is_authorized_for_cafe_action(
-            cafe, current_user, [Role.ADMIN, Role.VOLUNTEER]
+    item = await MenuService.get_item_by_id_and_cafe_id(id, cafe.id)
+    if not item:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=[{"msg": "An item with this ID does not exist."}],
         )
-    except ValueError as e:
-        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=str(e))
+
+    if data.category_id is not None:
+        category = await MenuService.get_category_by_id(cafe, data.category_id)
+        if not category:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail=[{"msg": "A category with this ID does not exist."}],
+            )
+
+    await CafeService.is_authorized_for_cafe_action(
+        cafe, current_user, [Role.ADMIN, Role.VOLUNTEER]
+    )
 
     try:
-        updated_item = await MenuService.update_item(item, data)
-        return updated_item
-    except Exception as e:
-        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail=str(e))
+        return await MenuService.update_item(item, data)
+    except DuplicateKeyError as e:
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail=[
+                {
+                    "msg": "An item with these fields already exists.",
+                    "fields": list(e.details.get("keyPattern", {}).keys()),
+                }
+            ],
+        )
 
 
 @menu_router.delete(
@@ -228,28 +316,24 @@ async def update_menu_item(
     },
 )
 async def delete_menu_item(
+    slug: str = Path(..., description="Slug of the cafe"),
     id: PydanticObjectId = Path(..., description="ID of the menu item"),
     current_user: User = Depends(get_current_user),
 ):
     """Delete a menu item. (`admin`)"""
-    item = await MenuService.get_item(id)
-    if not item:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Menu item not found."
-        )
-
-    cafe = await CafeService.get(item.cafe_id)
+    cafe = await CafeService.get(slug)
     if not cafe:
         raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Cafe not found"
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=[{"msg": "A cafe with this slug does not exist."}],
         )
 
-    try:
-        await CafeService.is_authorized_for_cafe_action(
-            cafe, current_user, [Role.ADMIN]
+    item = await MenuService.get_item_by_id_and_cafe_id(id, cafe.id)
+    if not item:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=[{"msg": "An item with this ID does not exist."}],
         )
-    except ValueError as e:
-        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=str(e))
 
+    await CafeService.is_authorized_for_cafe_action(cafe, current_user, [Role.ADMIN])
     await MenuService.delete_item(item)
-    return {"message": f"Item {id} has been successfully deleted."}
